+++ tests/testthat/test-cpu.R
@@ -1,7 +1,7 @@
 test_that("Test CPU", {
   skip_on_cran()
   cpu = get_cpu()
-  expect_equal(length(cpu), 2)
+  expect_equal(length(cpu), 3)
   expect_equal(anyNA(cpu), FALSE)
 }
 )