+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Version 0.2.0
   * More flexibilty in plot and datatable functions - you can now specify the test you want to compare.
+  * The number of cores returned by get_cpu().
 
 ## Version 0.1.9
   * Default log scale on y-axis (suggested by @eddelbuettel). Fixes #5.